Edit file workflow for creating a fork and proposing changes
When viewing a file that the user can't edit because they can't write to
the branch, the edit button is no longer disabled.

If no user fork of the repository exists, there is now a page create one.
It will automatically create a fork with a single branch matching the one
being viewed, and a unique repository name will be automatically picked.

If the fork exists, an message will explain at the top of the edit page
explaining that the changes will be applied to a branch in the fork. The
base repository branch will be pushed to a new branch to the fork, and
then the edits will be applied on top.

This all happens when accessing /_edit/, so that for example online
documentation can have an "edit this page" link to the base repository
that does the right thing.